# server/docs/swagger.yaml
definitions:
...
  orderManager.GoldGoods:
    properties:
      ID:
        description: 主键ID
        type: integer
      createId:
        description: 商品创建人
        type: integer
      createTime:
        description: 创建时间
        type: string
      createdAt:
        description: 创建时间
        type: string
      goodsName:
        description: 商品名称
        type: string
      goodsPrice:
        description: 商品价格
        type: integer
      goodsTypeId:
        description: 商品分类Id
        type: integer
      updateId:
        description: 商品更新人
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
  orderManager.GoldGoodsFile:
    properties:
      ID:
        description: 主键ID
        type: integer
      createId:
        description: 创建人id
        type: integer
      createTime:
        description: 创建时间
        type: string
      createdAt:
        description: 创建时间
        type: string
      fileName:
        description: 文件名
        type: string
      filePath:
        description: 文件路径
        type: string
      fileType:
        description: 文件类型
        type: boolean
      goodsId:
        description: 商品Id
        type: integer
      updateId:
        description: 更新人Id
        type: integer
      updateTime:
        description: 更新时间
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.AddGoldGoodsFileRequest:
    properties:
      ID:
        description: 主键ID
        type: integer
      createId:
        description: 创建人id
        type: integer
      createTime:
        description: 创建时间
        type: string
      fileName:
        description: 文件名
        type: string
      filePath:
        description: 文件路径
        type: string
      fileType:
        description: 文件类型
        type: boolean
      goodsId:
        description: 商品Id
        type: integer
      updateId:
        description: 更新人Id
        type: integer
      updateTime:
        description: 更新时间
        type: string
    type: object
  request.AddGoldGoodsRequest:
    properties:
      ID:
        description: 主键ID
        type: integer
      goldGoodsFileList:
        description: 上传文件路径
        items:
          $ref: '#/definitions/request.AddGoldGoodsFileRequest'
        type: array
      goodsName:
        description: 商品名称
        type: string
      goodsPrice:
        description: 商品价格
        type: integer
      goodsTypeId:
        description: 商品分类Id
        type: integer
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.5.9
paths:
...
  /goldGoods/AddGoldGoodsAndFiles:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建goldGoods表还有图片文件
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AddGoldGoodsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建goldGoods表还有图片文件
      tags:
      - GoldGoods
  /goldGoods/createGoldGoods:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建goldGoods表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldGoods'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建goldGoods表
      tags:
      - GoldGoods
  /goldGoods/deleteGoldGoods:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除goldGoods表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldGoods'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除goldGoods表
      tags:
      - GoldGoods
  /goldGoods/deleteGoldGoodsByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除goldGoods表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除goldGoods表
      tags:
      - GoldGoods
  /goldGoods/findGoldGoods:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 商品创建人
        in: query
        name: createId
        type: integer
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 商品名称
        in: query
        name: goodsName
        type: string
      - description: 商品价格
        in: query
        name: goodsPrice
        type: integer
      - description: 商品分类Id
        in: query
        name: goodsTypeId
        type: integer
      - description: 商品更新人
        in: query
        name: updateId
        type: integer
      - description: 更新时间
        in: query
        name: updateTime
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询goldGoods表
      tags:
      - GoldGoods
  /goldGoods/getGoldGoodsList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取goldGoods表列表
      tags:
      - GoldGoods
  /goldGoods/updateGoldGoods:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新goldGoods表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldGoods'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新goldGoods表
      tags:
      - GoldGoods
  /goldGoodsFile/createGoldGoodsFile:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建goldGoodsFile表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldGoodsFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建goldGoodsFile表
      tags:
      - GoldGoodsFile
  /goldGoodsFile/deleteGoldGoodsFile:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除goldGoodsFile表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldGoodsFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除goldGoodsFile表
      tags:
      - GoldGoodsFile
  /goldGoodsFile/deleteGoldGoodsFileByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除goldGoodsFile表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除goldGoodsFile表
      tags:
      - GoldGoodsFile
  /goldGoodsFile/findGoldGoodsFile:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 创建人id
        in: query
        name: createId
        type: integer
      - description: 创建时间
        in: query
        name: createTime
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 文件名
        in: query
        name: fileName
        type: string
      - description: 文件路径
        in: query
        name: filePath
        type: string
      - description: 文件类型
        in: query
        name: fileType
        type: boolean
      - description: 商品Id
        in: query
        name: goodsId
        type: integer
      - description: 更新人Id
        in: query
        name: updateId
        type: integer
      - description: 更新时间
        in: query
        name: updateTime
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询goldGoodsFile表
      tags:
      - GoldGoodsFile
  /goldGoodsFile/getGoldGoodsFileList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取goldGoodsFile表列表
      tags:
      - GoldGoodsFile
  /goldGoodsFile/updateGoldGoodsFile:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新goldGoodsFile表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/orderManager.GoldGoodsFile'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新goldGoodsFile表
      tags:
      - GoldGoodsFile
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"